Add files via upload
Our Fake Product Identification Through Barcode Using Blockchain Technology System – a Dotnet project proposes a form of identification of fake products using blockchain technology. In the proposed system, we are assigning a barcode to a particular product created by the manufacturer along with all the details of the product and the end customer can scan that barcode to get all information about that product. After scanning the barcode, the user can identify whether the product is real or fake